+++ components/Collections.vue
@@ -1,6 +1,7 @@
 <template>
 	<div class="vue-component collections">
 		<SearchableList :data="collections" identfierKey="id" summaryKey="title" :showSummaryOnExpand="false" :externalSearchTerm="searchTerm" :sort="sort" :allowExpand="allowExpand" :heading="heading">
+			<template #heading="scope"><slot name="heading" v-bind="scope" /></template>
 			<template #details="slot">
 				<Collection :collection="slot.item" :mapOptions="mapOptions">
 					<template #title><span class="hidden" /></template>